Trying to recover SPs on context change, implementing alpha memory layout

+++ Aufgabe1/vorgabe/include/scheduler.h
@@ -3,7 +3,7 @@
 #ifndef SCHEDULER_H
 #define SCHEDULER_H
 
-uint16_t rrSchedule(uint16_t, uint16_t);
+uint16_t rrSchedule(int , uint16_t);
 void adjustSleptTime();
 
 #endif //SCHEDULER_H
